+++ components/socials/social-join-dates.tsx
@@ -1,0 +1,20 @@
+import { SocialAccountResponse, SocialAccountWithoutAuthResponse } from '@/generated/graphql';
+import SocialJoinDate from './social-join-date';
+
+export default function SocialJoinDates({
+  socialAccounts,
+  logoSize,
+}: {
+  socialAccounts?: SocialAccountResponse[] | SocialAccountWithoutAuthResponse[];
+  logoSize: number,
+}) {
+  if (!socialAccounts) return null;
+
+  return (
+    <div className="flex flex-col justify-start">
+      {socialAccounts.map((socialAccount) => (
+        <SocialJoinDate socialAccount={socialAccount} logoSize={logoSize}/>
+      ))}
+    </div>
+  );
+}
